Split engine/engine_test.go into focused test files
Move TestProcessItem_* to process_item_test.go, TestPoll_* to poll_test.go,
TestAdvanceToNextStage_* to advance_stage_test.go, TestFormat*/TestCaptureGitMeta_*
to format_test.go, shared helpers to engine_helpers_test.go. engine_test.go retains
remaining tests.
