+++ engine/engine_test.go
@@ -4,122 +4,13 @@ import (
 	"context"
 	"fmt"
 	"os"
-	"os/exec"
-	"path/filepath"
-	"strings"
 	"sync"
 	"syscall"
 	"testing"
 	"time"
 
 	gh "github.com/verveguy/fabrik/github"
-	"github.com/verveguy/fabrik/stages"
 )
-
-func testStages() []*stages.Stage {
-	return []*stages.Stage{
-		{
-			Name:       "Research",
-			Order:      1,
-			Prompt:     "Do research",
-			Completion: stages.CompletionCriteria{Type: "claude"},
-		},
-		{
-			Name:       "Plan",
-			Order:      2,
-			Prompt:     "Make a plan",
-			Completion: stages.CompletionCriteria{Type: "claude"},
-		},
-		{
-			Name:       "Implement",
-			Order:      3,
-			Prompt:     "Implement it",
-			Completion: stages.CompletionCriteria{Type: "claude"},
-		},
-	}
-}
-
-func testEngine(client *mockGitHubClient, claude *mockClaudeInvoker) *Engine {
-	return NewWithDeps(
-		Config{
-			Owner:         "owner",
-			Repo:          "repo",
-			ProjectNum:    1,
-			User:          "testuser",
-			Token:         "token",
-			MaxConcurrent: 5,
-			Stages:        testStages(),
-		},
-		client,
-		claude,
-		NewWorktreeManager("/tmp/test-repo"),
-	)
-}
-
-func TestProcessItem_SkipsUnknownStage(t *testing.T) {
-	client := &mockGitHubClient{}
-	claude := &mockClaudeInvoker{}
-	eng := testEngine(client, claude)
-
-	board := &gh.ProjectBoard{ProjectID: "PVT_1"}
-	item := gh.ProjectItem{
-		Number: 1,
-		Title:  "Test",
-		Status: "Unknown Column",
-	}
-
-	err := eng.processItem(context.Background(), board, item)
-	if err != nil {
-		t.Fatalf("processItem: %v", err)
-	}
-	if len(claude.calls) != 0 {
-		t.Error("should not invoke claude for unknown stage")
-	}
-}
-
-func TestProcessItem_SkipsLockedByOther(t *testing.T) {
-	client := &mockGitHubClient{}
-	claude := &mockClaudeInvoker{}
-	eng := testEngine(client, claude)
-
-	board := &gh.ProjectBoard{ProjectID: "PVT_1"}
-	item := gh.ProjectItem{
-		Number: 1,
-		Title:  "Test",
-		Status: "Research",
-		Labels: []string{"fabrik:locked:otheruser"},
-	}
-
-	err := eng.processItem(context.Background(), board, item)
-	if err != nil {
-		t.Fatalf("processItem: %v", err)
-	}
-	if len(claude.calls) != 0 {
-		t.Error("should not invoke claude for item locked by another user")
-	}
-}
-
-func TestProcessItem_SkipsPaused(t *testing.T) {
-	client := &mockGitHubClient{}
-	claude := &mockClaudeInvoker{}
-	eng := testEngine(client, claude)
-
-	board := &gh.ProjectBoard{ProjectID: "PVT_1"}
-	item := gh.ProjectItem{
-		Number: 1,
-		Title:  "Test",
-		Status: "Research",
-		Labels: []string{"fabrik:paused"},
-	}
-
-	err := eng.processItem(context.Background(), board, item)
-	if err != nil {
-		t.Fatalf("processItem: %v", err)
-	}
-	if len(claude.calls) != 0 {
-		t.Error("should not invoke claude for paused item")
-	}
-}
 
 func TestItemNeedsWork_SkipsPaused(t *testing.T) {
 	eng := testEngine(&mockGitHubClient{}, &mockClaudeInvoker{})
@@ -152,83 +43,6 @@ func TestItemNeedsWork_SkipsPausedWithNewComments(t *testing.T) {
 	}
 }
 
-func TestProcessItem_AllowsOwnLock(t *testing.T) {
-	client := &mockGitHubClient{}
-	claude := &mockClaudeInvoker{
-		invokeFn: func(stage *stages.Stage, issue gh.ProjectItem, newComments []gh.Comment, resume bool, workDir string, modelOverride string) (string, bool, TokenUsage, error) {
-			return "output", false, TokenUsage{}, nil
-		},
-	}
-	eng := testEngine(client, claude)
-	// Need a real worktree manager for processItem — use mock that returns a temp dir
-	eng.worktrees = &WorktreeManager{baseDir: t.TempDir(), rootDir: t.TempDir()}
-
-	board := &gh.ProjectBoard{ProjectID: "PVT_1"}
-	item := gh.ProjectItem{
-		Number: 1,
-		Title:  "Test",
-		Status: "Research",
-		Labels: []string{"fabrik:locked:testuser"},
-	}
-
-	// processItem calls EnsureWorktree which needs git — skip worktree by mocking
-	// Instead, test that own lock doesn't cause skip by checking that we attempt to process
-	// We can't fully test processItem without git, so just test the lock check logic
-	err := eng.processItem(context.Background(), board, item)
-	// This will fail on EnsureWorktree since we don't have a real git repo,
-	// but the important thing is it didn't skip due to lock
-	if err != nil && !strings.Contains(err.Error(), "worktree") {
-		t.Fatalf("unexpected error: %v", err)
-	}
-}
-
-func TestProcessItem_SkipsCompleted(t *testing.T) {
-	client := &mockGitHubClient{}
-	claude := &mockClaudeInvoker{}
-	eng := testEngine(client, claude)
-
-	board := &gh.ProjectBoard{ProjectID: "PVT_1"}
-	item := gh.ProjectItem{
-		Number: 1,
-		Title:  "Test",
-		Status: "Research",
-		Labels: []string{"stage:Research:complete"},
-	}
-
-	err := eng.processItem(context.Background(), board, item)
-	if err != nil {
-		t.Fatalf("processItem: %v", err)
-	}
-	if len(claude.calls) != 0 {
-		t.Error("should not invoke claude for completed item")
-	}
-}
-
-func TestProcessItem_SkipsAlreadyProcessedNoNewComments(t *testing.T) {
-	client := &mockGitHubClient{}
-	claude := &mockClaudeInvoker{}
-	eng := testEngine(client, claude)
-	eng.cfg.PollSeconds = 100 // cooldown = 1000s — ensures recently-processed item stays in cooldown
-
-	// Mark as already processed
-	eng.processedSet["1-Research"] = time.Now()
-
-	board := &gh.ProjectBoard{ProjectID: "PVT_1"}
-	item := gh.ProjectItem{
-		Number: 1,
-		Title:  "Test",
-		Status: "Research",
-	}
-
-	err := eng.processItem(context.Background(), board, item)
-	if err != nil {
-		t.Fatalf("processItem: %v", err)
-	}
-	if len(claude.calls) != 0 {
-		t.Error("should not invoke claude when already processed and no new comments")
-	}
-}
-
 func TestFindNewComments(t *testing.T) {
 	eng := testEngine(&mockGitHubClient{}, &mockClaudeInvoker{})
 
@@ -258,191 +72,6 @@ func TestFindNewComments(t *testing.T) {
 	}
 }
 
-func TestAdvanceToNextStage_Success(t *testing.T) {
-	client := &mockGitHubClient{}
-	eng := testEngine(client, &mockClaudeInvoker{})
-	eng.statusField = &gh.StatusField{
-		FieldID: "FIELD_1",
-		Options: map[string]string{
-			"Research": "OPT_1",
-			"Plan":     "OPT_2",
-		},
-	}
-
-	board := &gh.ProjectBoard{ProjectID: "PVT_1"}
-	item := gh.ProjectItem{Number: 1, ItemID: "PVTI_1"}
-	stage := &stages.Stage{Name: "Research"}
-
-	err := eng.advanceToNextStage(board, item, stage)
-	if err != nil {
-		t.Fatalf("advanceToNextStage: %v", err)
-	}
-	if len(client.updateStatusCalls) != 1 {
-		t.Fatalf("expected 1 update call, got %d", len(client.updateStatusCalls))
-	}
-	call := client.updateStatusCalls[0]
-	if call.projectID != "PVT_1" || call.optionID != "OPT_2" {
-		t.Errorf("update call = %+v", call)
-	}
-}
-
-func TestAdvanceToNextStage_LastStage(t *testing.T) {
-	client := &mockGitHubClient{}
-	eng := testEngine(client, &mockClaudeInvoker{})
-	eng.statusField = &gh.StatusField{
-		FieldID: "FIELD_1",
-		Options: map[string]string{},
-	}
-
-	board := &gh.ProjectBoard{ProjectID: "PVT_1"}
-	item := gh.ProjectItem{Number: 1}
-	stage := &stages.Stage{Name: "Implement"} // last stage
-
-	err := eng.advanceToNextStage(board, item, stage)
-	if err != nil {
-		t.Fatalf("advanceToNextStage: %v", err)
-	}
-	if len(client.updateStatusCalls) != 0 {
-		t.Error("should not update status for last stage")
-	}
-}
-
-func TestAdvanceToNextStage_NoStatusField(t *testing.T) {
-	eng := testEngine(&mockGitHubClient{}, &mockClaudeInvoker{})
-	// statusField is nil
-
-	board := &gh.ProjectBoard{ProjectID: "PVT_1"}
-	item := gh.ProjectItem{Number: 1}
-	stage := &stages.Stage{Name: "Research"}
-
-	err := eng.advanceToNextStage(board, item, stage)
-	if err == nil {
-		t.Fatal("expected error when statusField is nil")
-	}
-}
-
-func TestAdvanceToNextStage_MissingOption(t *testing.T) {
-	eng := testEngine(&mockGitHubClient{}, &mockClaudeInvoker{})
-	eng.statusField = &gh.StatusField{
-		FieldID: "FIELD_1",
-		Options: map[string]string{
-			"Research": "OPT_1",
-			// Plan option is missing
-		},
-	}
-
-	board := &gh.ProjectBoard{ProjectID: "PVT_1"}
-	item := gh.ProjectItem{Number: 1}
-	stage := &stages.Stage{Name: "Research"}
-
-	err := eng.advanceToNextStage(board, item, stage)
-	if err == nil {
-		t.Fatal("expected error for missing option")
-	}
-}
-
-func TestFormatOutputComment(t *testing.T) {
-	comment := formatOutputComment("Research", "Hello world", "", "main", "abc12345", "", "2026-04-01 14:30 UTC")
-	if !strings.Contains(comment, "🏭 **Fabrik — stage: Research**") {
-		t.Errorf("comment = %q", comment)
-	}
-	if !strings.Contains(comment, "Hello world") {
-		t.Error("comment missing output")
-	}
-	if !strings.Contains(comment, "*branch: main | commit: abc12345 | 2026-04-01 14:30 UTC*") {
-		t.Errorf("comment missing metadata line: %q", comment)
-	}
-}
-
-func TestFormatOutputComment_WithMainSHA(t *testing.T) {
-	comment := formatOutputComment("Research", "Hello", "", "main", "abc12345", "def56789", "2026-04-01 14:30 UTC")
-	expected := "*branch: main | commit: abc12345 | main: def56789 | 2026-04-01 14:30 UTC*"
-	if !strings.Contains(comment, expected) {
-		t.Errorf("comment missing main SHA metadata line.\ngot: %q\nwant substring: %q", comment, expected)
-	}
-}
-
-func TestFormatOutputComment_Truncation(t *testing.T) {
-	longOutput := strings.Repeat("x", 70000)
-	comment := formatOutputComment("Test", longOutput, "", "main", "abc12345", "", "2026-04-01 14:30 UTC")
-	if len(comment) > 61000 {
-		t.Errorf("comment should be truncated, len = %d", len(comment))
-	}
-	if !strings.Contains(comment, "... (truncated)") {
-		t.Error("truncated comment missing truncation notice")
-	}
-}
-
-func TestFormatOutputComment_FooterSurvivesTruncation(t *testing.T) {
-	// Output exceeds the limit; footer must appear after the truncation notice, not be cut off.
-	longOutput := strings.Repeat("x", 65000)
-	footer := "\n\n---\nUsed 30/30 turns, 47k input / 8k output tokens. Stage incomplete."
-	comment := formatOutputComment("Test", longOutput, footer, "main", "abc12345", "", "2026-04-01 14:30 UTC")
-	if !strings.Contains(comment, "... (truncated)") {
-		t.Error("expected truncation notice")
-	}
-	if !strings.Contains(comment, "30/30 turns") {
-		t.Error("stats footer must appear after truncation, not be cut off")
-	}
-}
-
-func TestFormatPRSummaryComment(t *testing.T) {
-	output := "FABRIK_SUMMARY_BEGIN\nDid some work.\nFABRIK_SUMMARY_END\n"
-	comment := formatPRSummaryComment("Plan", 42, output, "fabrik/issue-5", "deadbeef", "", "2026-04-01 14:30 UTC")
-	if !strings.Contains(comment, "🏭 **Fabrik — stage: Plan**") {
-		t.Errorf("missing header: %q", comment)
-	}
-	if !strings.Contains(comment, "*branch: fabrik/issue-5 | commit: deadbeef | 2026-04-01 14:30 UTC*") {
-		t.Errorf("missing metadata line: %q", comment)
-	}
-	if !strings.Contains(comment, "PR #42") {
-		t.Errorf("missing PR reference: %q", comment)
-	}
-}
-
-func TestFormatPRSummaryComment_WithMainSHA(t *testing.T) {
-	output := "FABRIK_SUMMARY_BEGIN\nDid some work.\nFABRIK_SUMMARY_END\n"
-	comment := formatPRSummaryComment("Plan", 42, output, "fabrik/issue-5", "deadbeef", "aaa11111", "2026-04-01 14:30 UTC")
-	expected := "*branch: fabrik/issue-5 | commit: deadbeef | main: aaa11111 | 2026-04-01 14:30 UTC*"
-	if !strings.Contains(comment, expected) {
-		t.Errorf("missing main SHA metadata.\ngot: %q\nwant substring: %q", comment, expected)
-	}
-}
-
-func TestCaptureGitMeta_EmptyWorkDir(t *testing.T) {
-	branch, commit, mainSHA, timestamp := captureGitMeta("", "")
-	if branch != "unknown" {
-		t.Errorf("expected branch=unknown, got %q", branch)
-	}
-	if commit != "unknown" {
-		t.Errorf("expected commit=unknown, got %q", commit)
-	}
-	if mainSHA != "" {
-		t.Errorf("expected empty mainSHA, got %q", mainSHA)
-	}
-	if timestamp == "" {
-		t.Error("expected non-empty timestamp")
-	}
-}
-
-func TestCaptureGitMeta_ValidDir(t *testing.T) {
-	// Use the current repo root — it definitely has commits and a branch
-	repoRoot, err := gitToplevel()
-	if err != nil {
-		t.Skip("not in a git repo")
-	}
-	branch, commit, _, timestamp := captureGitMeta(repoRoot, "main")
-	if branch == "unknown" {
-		t.Errorf("expected real branch, got %q", branch)
-	}
-	if commit == "unknown" || len(commit) != 8 {
-		t.Errorf("expected 8-char commit SHA, got %q", commit)
-	}
-	if timestamp == "" {
-		t.Error("expected non-empty timestamp")
-	}
-}
-
 func TestMapKeys(t *testing.T) {
 	m := map[string]string{
 		"a": "1",
@@ -462,49 +91,6 @@ func TestMapKeys(t *testing.T) {
 		if !found[expected] {
 			t.Errorf("missing key %q", expected)
 		}
-	}
-}
-
-func TestPoll_FetchesBoardAndProcessesItems(t *testing.T) {
-	client := &mockGitHubClient{
-		fetchProjectBoardFn: func(owner, repo string, projectNum int) (*gh.ProjectBoard, error) {
-			return &gh.ProjectBoard{
-				ProjectID: "PVT_1",
-				Items: []gh.ProjectItem{
-					{Number: 1, Title: "Test", Status: "Unknown"},
-				},
-			}, nil
-		},
-		fetchStatusFieldFn: func(projectID string) (*gh.StatusField, error) {
-			return &gh.StatusField{
-				FieldID: "F1",
-				Options: map[string]string{"Research": "OPT_1"},
-			}, nil
-		},
-	}
-	eng := testEngine(client, &mockClaudeInvoker{})
-
-	err := eng.poll(context.Background())
-	if err != nil {
-		t.Fatalf("poll: %v", err)
-	}
-	// Status field should be fetched
-	if eng.statusField == nil {
-		t.Error("statusField should be set after poll")
-	}
-}
-
-func TestPoll_Error(t *testing.T) {
-	client := &mockGitHubClient{
-		fetchProjectBoardFn: func(owner, repo string, projectNum int) (*gh.ProjectBoard, error) {
-			return nil, fmt.Errorf("network error")
-		},
-	}
-	eng := testEngine(client, &mockClaudeInvoker{})
-
-	err := eng.poll(context.Background())
-	if err == nil {
-		t.Fatal("expected error from poll")
 	}
 }
 
@@ -531,545 +117,6 @@ func TestGitToplevel(t *testing.T) {
 	}
 	if dir == "" {
 		t.Error("gitToplevel returned empty string")
-	}
-}
-
-func TestProcessItem_FullHappyPath(t *testing.T) {
-	skipIfNoGit(t)
-	repoDir := initBareRepo(t)
-	wm := NewWorktreeManager(repoDir)
-
-	client := &mockGitHubClient{}
-	claude := &mockClaudeInvoker{
-		invokeFn: func(stage *stages.Stage, issue gh.ProjectItem, newComments []gh.Comment, resume bool, workDir string, modelOverride string) (string, bool, TokenUsage, error) {
-			return "Claude output here\nFABRIK_STAGE_COMPLETE\n", true, TokenUsage{}, nil
-		},
-	}
-
-	eng := NewWithDeps(
-		Config{
-			Owner:      "owner",
-			Repo:       "repo",
-			ProjectNum: 1,
-			User:       "testuser",
-			Token:      "token",
-			Stages:     testStages(),
-		},
-		client,
-		claude,
-		wm,
-	)
-
-	board := &gh.ProjectBoard{ProjectID: "PVT_1"}
-	item := gh.ProjectItem{
-		Number: 1,
-		Title:  "Test Issue",
-		Status: "Research",
-		ItemID: "PVTI_1",
-	}
-
-	err := eng.processItem(context.Background(), board, item)
-	if err != nil {
-		t.Fatalf("processItem: %v", err)
-	}
-
-	// Should have locked the issue
-	if len(client.addLabelCalls) < 1 {
-		t.Fatal("expected lock label call")
-	}
-	if client.addLabelCalls[0].labelName != "fabrik:locked:testuser" {
-		t.Errorf("lock label = %q", client.addLabelCalls[0].labelName)
-	}
-
-	// Lock label is released when stage completes (completed=true → releaseLock() called).
-	// When not completed, the lock persists through cooldown so other instances don't
-	// pick up the issue — see "Keep lock and in_progress labels through cooldown retries".
-	foundLockRemoval := false
-	for _, call := range client.removeLabelCalls {
-		if call.labelName == "fabrik:locked:testuser" {
-			foundLockRemoval = true
-		}
-	}
-	if !foundLockRemoval {
-		t.Error("expected lock label to be removed after stage completes")
-	}
-
-	// Should have invoked Claude
-	if len(claude.calls) != 1 {
-		t.Fatalf("expected 1 claude call, got %d", len(claude.calls))
-	}
-	if claude.calls[0].stageName != "Research" {
-		t.Errorf("stage = %q", claude.calls[0].stageName)
-	}
-
-	// Should have posted comment
-	if len(client.addCommentCalls) != 1 {
-		t.Fatalf("expected 1 comment call, got %d", len(client.addCommentCalls))
-	}
-	if !strings.Contains(client.addCommentCalls[0].body, "Claude output here") {
-		t.Errorf("comment = %q", client.addCommentCalls[0].body)
-	}
-}
-
-func TestProcessItem_AccumulatesTokenUsage(t *testing.T) {
-	skipIfNoGit(t)
-	repoDir := initBareRepo(t)
-	wm := NewWorktreeManager(repoDir)
-
-	want := TokenUsage{InputTokens: 100, OutputTokens: 50, CacheCreationTokens: 10, CacheReadTokens: 5, CostUSD: 0.0042}
-	client := &mockGitHubClient{}
-	claude := &mockClaudeInvoker{
-		invokeFn: func(stage *stages.Stage, issue gh.ProjectItem, newComments []gh.Comment, resume bool, workDir string, modelOverride string) (string, bool, TokenUsage, error) {
-			return "FABRIK_STAGE_COMPLETE", true, want, nil
-		},
-	}
-
-	eng := NewWithDeps(
-		Config{Owner: "o", Repo: "r", User: "u", Token: "t", Stages: testStages()},
-		client, claude, wm,
-	)
-
-	board := &gh.ProjectBoard{ProjectID: "PVT_1"}
-	item := gh.ProjectItem{Number: 200, Title: "Token Test", Status: "Research", ItemID: "PVTI_200"}
-
-	if err := eng.processItem(context.Background(), board, item); err != nil {
-		t.Fatalf("processItem: %v", err)
-	}
-
-	eng.mu.Lock()
-	got := eng.totalTokens
-	eng.mu.Unlock()
-
-	if got.InputTokens != want.InputTokens {
-		t.Errorf("totalTokens.InputTokens = %d, want %d", got.InputTokens, want.InputTokens)
-	}
-	if got.OutputTokens != want.OutputTokens {
-		t.Errorf("totalTokens.OutputTokens = %d, want %d", got.OutputTokens, want.OutputTokens)
-	}
-	if got.CacheCreationTokens != want.CacheCreationTokens {
-		t.Errorf("totalTokens.CacheCreationTokens = %d, want %d", got.CacheCreationTokens, want.CacheCreationTokens)
-	}
-	if got.CacheReadTokens != want.CacheReadTokens {
-		t.Errorf("totalTokens.CacheReadTokens = %d, want %d", got.CacheReadTokens, want.CacheReadTokens)
-	}
-	if diff := got.CostUSD - want.CostUSD; diff > 1e-9 || diff < -1e-9 {
-		t.Errorf("totalTokens.CostUSD = %f, want ~%f", got.CostUSD, want.CostUSD)
-	}
-}
-
-func TestProcessItem_CompletionWithAutoAdvance(t *testing.T) {
-	skipIfNoGit(t)
-	repoDir := initBareRepo(t)
-	wm := NewWorktreeManager(repoDir)
-
-	client := &mockGitHubClient{}
-	claude := &mockClaudeInvoker{
-		invokeFn: func(stage *stages.Stage, issue gh.ProjectItem, newComments []gh.Comment, resume bool, workDir string, modelOverride string) (string, bool, TokenUsage, error) {
-			return "Done\nFABRIK_STAGE_COMPLETE", true, TokenUsage{}, nil
-		},
-	}
-
-	eng := NewWithDeps(
-		Config{
-			Owner:      "owner",
-			Repo:       "repo",
-			ProjectNum: 1,
-			User:       "testuser",
-			Token:      "token",
-			Yolo:       true,
-			Stages:     testStages(),
-		},
-		client,
-		claude,
-		wm,
-	)
-	eng.statusField = &gh.StatusField{
-		FieldID: "F1",
-		Options: map[string]string{
-			"Research": "OPT_1",
-			"Plan":     "OPT_2",
-		},
-	}
-
-	board := &gh.ProjectBoard{ProjectID: "PVT_1"}
-	item := gh.ProjectItem{
-		Number: 2,
-		Title:  "Auto advance test",
-		Status: "Research",
-		ItemID: "PVTI_2",
-	}
-
-	err := eng.processItem(context.Background(), board, item)
-	if err != nil {
-		t.Fatalf("processItem: %v", err)
-	}
-
-	// Should have added completion label
-	foundComplete := false
-	for _, call := range client.addLabelCalls {
-		if call.labelName == "stage:Research:complete" {
-			foundComplete = true
-		}
-	}
-	if !foundComplete {
-		t.Error("expected completion label to be added")
-	}
-
-	// Should have removed the lock label after processing completes
-	foundLockRemoval := false
-	for _, call := range client.removeLabelCalls {
-		if call.labelName == "fabrik:locked:testuser" {
-			foundLockRemoval = true
-		}
-	}
-	if !foundLockRemoval {
-		t.Error("expected lock label to be removed after processItem completes")
-	}
-
-	// Should have advanced to next stage
-	if len(client.updateStatusCalls) != 1 {
-		t.Fatalf("expected 1 status update, got %d", len(client.updateStatusCalls))
-	}
-	if client.updateStatusCalls[0].optionID != "OPT_2" {
-		t.Errorf("advanced to option = %q, want OPT_2", client.updateStatusCalls[0].optionID)
-	}
-}
-
-func TestProcessItem_CompletionNoAutoAdvance(t *testing.T) {
-	skipIfNoGit(t)
-	repoDir := initBareRepo(t)
-	wm := NewWorktreeManager(repoDir)
-
-	client := &mockGitHubClient{}
-	claude := &mockClaudeInvoker{
-		invokeFn: func(stage *stages.Stage, issue gh.ProjectItem, newComments []gh.Comment, resume bool, workDir string, modelOverride string) (string, bool, TokenUsage, error) {
-			return "Done", true, TokenUsage{}, nil
-		},
-	}
-
-	eng := NewWithDeps(
-		Config{
-			Owner:      "owner",
-			Repo:       "repo",
-			ProjectNum: 1,
-			User:       "testuser",
-			Token:      "token",
-			Yolo:       false, // no auto-advance
-			Stages:     testStages(),
-		},
-		client,
-		claude,
-		wm,
-	)
-
-	board := &gh.ProjectBoard{ProjectID: "PVT_1"}
-	item := gh.ProjectItem{Number: 3, Title: "No advance", Status: "Research", ItemID: "PVTI_3"}
-
-	err := eng.processItem(context.Background(), board, item)
-	if err != nil {
-		t.Fatalf("processItem: %v", err)
-	}
-
-	// Should NOT have advanced
-	if len(client.updateStatusCalls) != 0 {
-		t.Error("should not advance when yolo=false")
-	}
-}
-
-func TestProcessItem_StageAutoAdvanceOverride(t *testing.T) {
-	skipIfNoGit(t)
-	repoDir := initBareRepo(t)
-	wm := NewWorktreeManager(repoDir)
-
-	client := &mockGitHubClient{}
-	claude := &mockClaudeInvoker{
-		invokeFn: func(stage *stages.Stage, issue gh.ProjectItem, newComments []gh.Comment, resume bool, workDir string, modelOverride string) (string, bool, TokenUsage, error) {
-			return "Done", true, TokenUsage{}, nil
-		},
-	}
-
-	autoAdvance := true
-	stgs := []*stages.Stage{
-		{Name: "Research", Order: 1, Prompt: "p", Completion: stages.CompletionCriteria{Type: "claude"}, AutoAdvance: &autoAdvance},
-		{Name: "Plan", Order: 2, Prompt: "p", Completion: stages.CompletionCriteria{Type: "claude"}},
-	}
-
-	eng := NewWithDeps(
-		Config{
-			Owner:  "owner",
-			Repo:   "repo",
-			User:   "testuser",
-			Token:  "token",
-			Yolo:   false, // global is false
-			Stages: stgs,
-		},
-		client,
-		claude,
-		wm,
-	)
-	eng.statusField = &gh.StatusField{
-		FieldID: "F1",
-		Options: map[string]string{"Plan": "OPT_2"},
-	}
-
-	board := &gh.ProjectBoard{ProjectID: "PVT_1"}
-	item := gh.ProjectItem{Number: 4, Title: "Override", Status: "Research", ItemID: "PVTI_4"}
-
-	err := eng.processItem(context.Background(), board, item)
-	if err != nil {
-		t.Fatalf("processItem: %v", err)
-	}
-
-	// Should advance due to stage-level override
-	if len(client.updateStatusCalls) != 1 {
-		t.Error("expected advance due to stage AutoAdvance override")
-	}
-}
-
-func TestProcessItem_EmptyOutput(t *testing.T) {
-	skipIfNoGit(t)
-	repoDir := initBareRepo(t)
-	wm := NewWorktreeManager(repoDir)
-
-	client := &mockGitHubClient{}
-	claude := &mockClaudeInvoker{
-		invokeFn: func(stage *stages.Stage, issue gh.ProjectItem, newComments []gh.Comment, resume bool, workDir string, modelOverride string) (string, bool, TokenUsage, error) {
-			return "", false, TokenUsage{}, nil
-		},
-	}
-
-	eng := NewWithDeps(
-		Config{Owner: "o", Repo: "r", User: "u", Token: "t", Stages: testStages()},
-		client, claude, wm,
-	)
-
-	board := &gh.ProjectBoard{ProjectID: "PVT_1"}
-	item := gh.ProjectItem{Number: 5, Title: "Empty", Status: "Research", ItemID: "PVTI_5"}
-
-	err := eng.processItem(context.Background(), board, item)
-	if err != nil {
-		t.Fatalf("processItem: %v", err)
-	}
-
-	// Should NOT post comment when output is empty
-	if len(client.addCommentCalls) != 0 {
-		t.Error("should not post comment for empty output")
-	}
-}
-
-func TestProcessItem_ClaudeError(t *testing.T) {
-	skipIfNoGit(t)
-	repoDir := initBareRepo(t)
-	wm := NewWorktreeManager(repoDir)
-
-	client := &mockGitHubClient{}
-	claude := &mockClaudeInvoker{
-		invokeFn: func(stage *stages.Stage, issue gh.ProjectItem, newComments []gh.Comment, resume bool, workDir string, modelOverride string) (string, bool, TokenUsage, error) {
-			// Simulate a start failure: binary not found (*exec.Error)
-			cmd := exec.Command("this-binary-does-not-exist-fabrik-test")
-			_, startErr := cmd.Output()
-			return "partial output", false, TokenUsage{}, startErr
-		},
-	}
-
-	eng := NewWithDeps(
-		Config{Owner: "o", Repo: "r", User: "u", Token: "t", Stages: testStages()},
-		client, claude, wm,
-	)
-
-	board := &gh.ProjectBoard{ProjectID: "PVT_1"}
-	item := gh.ProjectItem{Number: 6, Title: "Error", Status: "Research", ItemID: "PVTI_6"}
-
-	// Should not return error — claude errors are logged, not fatal
-	err := eng.processItem(context.Background(), board, item)
-	if err != nil {
-		t.Fatalf("processItem: %v", err)
-	}
-
-	// Should still post partial output
-	if len(client.addCommentCalls) != 1 {
-		t.Fatalf("expected 1 comment with partial output, got %d", len(client.addCommentCalls))
-	}
-
-	// A start-failure (*exec.Error / binary not found) — processedSet must NOT be updated
-	itemKey := fmt.Sprintf("%d-%s", 6, "Research")
-	eng.mu.Lock()
-	_, recorded := eng.processedSet[itemKey]
-	eng.mu.Unlock()
-	if recorded {
-		t.Error("processedSet should NOT be updated on a start-failure error")
-	}
-}
-
-func TestProcessItem_ClaudeExitError(t *testing.T) {
-	skipIfNoGit(t)
-	repoDir := initBareRepo(t)
-	wm := NewWorktreeManager(repoDir)
-
-	client := &mockGitHubClient{}
-	claude := &mockClaudeInvoker{
-		invokeFn: func(stage *stages.Stage, issue gh.ProjectItem, newComments []gh.Comment, resume bool, workDir string, modelOverride string) (string, bool, TokenUsage, error) {
-			// Simulate Claude running and exiting non-zero (wrapped *exec.ExitError)
-			cmd := exec.Command("git", "definitely-invalid-arg")
-			runErr := cmd.Run()
-			return "some output", false, TokenUsage{}, fmt.Errorf("claude exited with error: %w", runErr)
-		},
-	}
-
-	eng := NewWithDeps(
-		Config{Owner: "o", Repo: "r", User: "u", Token: "t", Stages: testStages()},
-		client, claude, wm,
-	)
-
-	board := &gh.ProjectBoard{ProjectID: "PVT_1"}
-	item := gh.ProjectItem{Number: 7, Title: "ExitError", Status: "Research", ItemID: "PVTI_7"}
-
-	err := eng.processItem(context.Background(), board, item)
-	if err != nil {
-		t.Fatalf("processItem: %v", err)
-	}
-
-	// An *exec.ExitError means Claude ran — processedSet MUST be updated (cooldown applies)
-	itemKey := fmt.Sprintf("%d-%s", 7, "Research")
-	eng.mu.Lock()
-	_, recorded := eng.processedSet[itemKey]
-	eng.mu.Unlock()
-	if !recorded {
-		t.Error("processedSet should be updated when Claude ran and exited non-zero")
-	}
-}
-
-func TestProcessItem_ResumeOnReprocess(t *testing.T) {
-	skipIfNoGit(t)
-	repoDir := initBareRepo(t)
-	wm := NewWorktreeManager(repoDir)
-
-	client := &mockGitHubClient{}
-	claude := &mockClaudeInvoker{
-		invokeFn: func(stage *stages.Stage, issue gh.ProjectItem, newComments []gh.Comment, resume bool, workDir string, modelOverride string) (string, bool, TokenUsage, error) {
-			return "output", false, TokenUsage{}, nil
-		},
-	}
-
-	eng := NewWithDeps(
-		Config{Owner: "o", Repo: "r", User: "u", Token: "t", Stages: testStages()},
-		client, claude, wm,
-	)
-
-	board := &gh.ProjectBoard{ProjectID: "PVT_1"}
-	item := gh.ProjectItem{
-		Number: 7,
-		Title:  "Resume test",
-		Status: "Research",
-		ItemID: "PVTI_7",
-		// No comments — both calls go through the stage invocation path (e.claude.Invoke).
-		// processComments uses InvokeClaudeForComments (global), not the mock.
-	}
-
-	// First call — not yet in processedSet, resume=false
-	eng.processItem(context.Background(), board, item)
-
-	// Second call — PollSeconds=0 means cooldown=0, so item is retried with resume=true
-	eng.processItem(context.Background(), board, item)
-
-	if len(claude.calls) != 2 {
-		t.Fatalf("expected 2 claude calls, got %d", len(claude.calls))
-	}
-	if claude.calls[0].resume != false {
-		t.Error("first call should not resume")
-	}
-	if claude.calls[1].resume != true {
-		t.Error("second call should resume")
-	}
-}
-
-func TestPoll_StatusFieldFetchError(t *testing.T) {
-	client := &mockGitHubClient{
-		fetchProjectBoardFn: func(owner, repo string, projectNum int) (*gh.ProjectBoard, error) {
-			return &gh.ProjectBoard{ProjectID: "PVT_1", Items: nil}, nil
-		},
-		fetchStatusFieldFn: func(projectID string) (*gh.StatusField, error) {
-			return nil, fmt.Errorf("status field error")
-		},
-	}
-	eng := testEngine(client, &mockClaudeInvoker{})
-
-	// Should not error — status field failure is a warning
-	err := eng.poll(context.Background())
-	if err != nil {
-		t.Fatalf("poll: %v", err)
-	}
-	if eng.statusField != nil {
-		t.Error("statusField should remain nil on fetch error")
-	}
-}
-
-func TestPoll_StatusFieldAlreadySet(t *testing.T) {
-	client := &mockGitHubClient{
-		fetchProjectBoardFn: func(owner, repo string, projectNum int) (*gh.ProjectBoard, error) {
-			return &gh.ProjectBoard{ProjectID: "PVT_1"}, nil
-		},
-		fetchStatusFieldFn: func(projectID string) (*gh.StatusField, error) {
-			t.Error("should not fetch status field again")
-			return nil, nil
-		},
-	}
-	eng := testEngine(client, &mockClaudeInvoker{})
-	eng.statusField = &gh.StatusField{FieldID: "already-set"}
-
-	eng.poll(context.Background())
-}
-
-func TestPoll_EmptyProjectID(t *testing.T) {
-	client := &mockGitHubClient{
-		fetchProjectBoardFn: func(owner, repo string, projectNum int) (*gh.ProjectBoard, error) {
-			return &gh.ProjectBoard{ProjectID: ""}, nil
-		},
-		fetchStatusFieldFn: func(projectID string) (*gh.StatusField, error) {
-			t.Error("should not fetch status field when projectID is empty")
-			return nil, nil
-		},
-	}
-	eng := testEngine(client, &mockClaudeInvoker{})
-
-	eng.poll(context.Background())
-}
-
-func TestPoll_RateLimitLogging(t *testing.T) {
-	resetTime := time.Now().Add(time.Hour)
-	client := &mockGitHubClient{
-		fetchProjectBoardFn: func(owner, repo string, projectNum int) (*gh.ProjectBoard, error) {
-			return &gh.ProjectBoard{ProjectID: "PVT_1"}, nil
-		},
-		rateLimitStatsFn: func() (gh.RateLimitStats, gh.RateLimitStats) {
-			rest := gh.RateLimitStats{Limit: 5000, Remaining: 4800, Used: 200, Reset: resetTime}
-			gql := gh.RateLimitStats{Limit: 5000, Remaining: 4950, Used: 50, Reset: resetTime}
-			return rest, gql
-		},
-	}
-	eng := testEngine(client, &mockClaudeInvoker{})
-
-	// poll() must succeed and not panic when rate limit stats are non-zero.
-	if err := eng.poll(context.Background()); err != nil {
-		t.Fatalf("poll: %v", err)
-	}
-}
-
-func TestPoll_RateLimitLogging_ZeroReset(t *testing.T) {
-	// Verify poll() handles a zero Reset (header absent) gracefully — no panic, no "00:00 UTC".
-	client := &mockGitHubClient{
-		fetchProjectBoardFn: func(owner, repo string, projectNum int) (*gh.ProjectBoard, error) {
-			return &gh.ProjectBoard{ProjectID: "PVT_1"}, nil
-		},
-		rateLimitStatsFn: func() (gh.RateLimitStats, gh.RateLimitStats) {
-			rest := gh.RateLimitStats{Limit: 60, Remaining: 0} // Reset is zero
-			return rest, gh.RateLimitStats{}
-		},
-	}
-	eng := testEngine(client, &mockClaudeInvoker{})
-
-	if err := eng.poll(context.Background()); err != nil {
-		t.Fatalf("poll: %v", err)
 	}
 }
 
@@ -1128,74 +175,6 @@ func TestRun_ShutdownOnSignal(t *testing.T) {
 	}
 }
 
-func TestProcessItem_LabelAndCommentErrors(t *testing.T) {
-	skipIfNoGit(t)
-	repoDir := initBareRepo(t)
-	wm := NewWorktreeManager(repoDir)
-
-	client := &mockGitHubClient{
-		addLabelToIssueFn: func(owner, repo string, issueNumber int, labelName string) error {
-			return fmt.Errorf("label error")
-		},
-		addCommentFn: func(owner, repo string, issueNumber int, body string) error {
-			return fmt.Errorf("comment error")
-		},
-	}
-	claude := &mockClaudeInvoker{
-		invokeFn: func(stage *stages.Stage, issue gh.ProjectItem, newComments []gh.Comment, resume bool, workDir string, modelOverride string) (string, bool, TokenUsage, error) {
-			return "output", true, TokenUsage{}, nil
-		},
-	}
-
-	eng := NewWithDeps(
-		Config{Owner: "o", Repo: "r", User: "u", Token: "t", Stages: testStages()},
-		client, claude, wm,
-	)
-
-	board := &gh.ProjectBoard{ProjectID: "PVT_1"}
-	item := gh.ProjectItem{Number: 8, Title: "Errors", Status: "Research", ItemID: "PVTI_8"}
-
-	// Should not return error — label/comment errors are logged, not fatal
-	err := eng.processItem(context.Background(), board, item)
-	if err != nil {
-		t.Fatalf("processItem: %v", err)
-	}
-}
-
-func TestPoll_ProcessItemError(t *testing.T) {
-	client := &mockGitHubClient{
-		fetchProjectBoardFn: func(owner, repo string, projectNum int) (*gh.ProjectBoard, error) {
-			return &gh.ProjectBoard{
-				ProjectID: "PVT_1",
-				Items: []gh.ProjectItem{
-					{Number: 1, Title: "Test", Status: "Research", ItemID: "PVTI_1"},
-				},
-			}, nil
-		},
-		fetchStatusFieldFn: func(projectID string) (*gh.StatusField, error) {
-			return &gh.StatusField{FieldID: "F1", Options: map[string]string{}}, nil
-		},
-	}
-	claude := &mockClaudeInvoker{
-		invokeFn: func(stage *stages.Stage, issue gh.ProjectItem, newComments []gh.Comment, resume bool, workDir string, modelOverride string) (string, bool, TokenUsage, error) {
-			return "", false, TokenUsage{}, nil
-		},
-	}
-
-	eng := NewWithDeps(
-		Config{Owner: "o", Repo: "r", User: "u", Token: "t", Stages: testStages()},
-		client, claude, NewWorktreeManager("/nonexistent"),
-	)
-
-	// poll should not return error even when processItem fails
-	err := eng.poll(context.Background())
-	if err != nil {
-		t.Fatalf("poll should not error from processItem failures: %v", err)
-	}
-}
-
-// TestProcessedSetConcurrency verifies that concurrent access to processedSet
-// via the mutex-protected methods does not cause data races.
 func TestProcessedSetConcurrency(t *testing.T) {
 	e := &Engine{
 		cfg:          Config{User: "testuser"},
@@ -1500,284 +479,6 @@ func TestExtractModelOverrideWarnsOnMultiple(t *testing.T) {
 	}
 }
 
-func TestProcessItem_EscalatesAtMaxRetries(t *testing.T) {
-	skipIfNoGit(t)
-	repoDir := initBareRepo(t)
-	wm := NewWorktreeManager(repoDir)
-
-	client := &mockGitHubClient{}
-	claude := &mockClaudeInvoker{
-		invokeFn: func(stage *stages.Stage, issue gh.ProjectItem, newComments []gh.Comment, resume bool, workDir string, modelOverride string) (string, bool, TokenUsage, error) {
-			return "partial output", false, TokenUsage{}, nil // never completes
-		},
-	}
-
-	eng := NewWithDeps(
-		Config{
-			Owner:      "owner",
-			Repo:       "repo",
-			ProjectNum: 1,
-			User:       "testuser",
-			Token:      "token",
-			MaxRetries: 2,
-			Stages:     testStages(),
-		},
-		client,
-		claude,
-		wm,
-	)
-
-	board := &gh.ProjectBoard{ProjectID: "PVT_1"}
-	item := gh.ProjectItem{Number: 10, Title: "Escalate test", Status: "Research", ItemID: "PVTI_10"}
-
-	// PollSeconds=0 makes cooldown=0, so both calls reach Claude without waiting.
-	// First attempt — retry count becomes 1, no escalation yet
-	if err := eng.processItem(context.Background(), board, item); err != nil {
-		t.Fatalf("processItem (first call): %v", err)
-	}
-	foundPaused := false
-	for _, call := range client.addLabelCalls {
-		if call.labelName == "fabrik:paused" {
-			foundPaused = true
-		}
-	}
-	if foundPaused {
-		t.Error("should not escalate after first failure")
-	}
-
-	// Second attempt — retry count becomes 2, should escalate
-	if err := eng.processItem(context.Background(), board, item); err != nil {
-		t.Fatalf("processItem (second call): %v", err)
-	}
-
-	foundPaused = false
-	foundFailed := false
-	for _, call := range client.addLabelCalls {
-		if call.labelName == "fabrik:paused" {
-			foundPaused = true
-		}
-		if call.labelName == "stage:Research:failed" {
-			foundFailed = true
-		}
-	}
-	if !foundPaused {
-		t.Error("expected fabrik:paused label after max retries")
-	}
-	if !foundFailed {
-		t.Error("expected stage:Research:failed label after max retries")
-	}
-
-	// Should have posted an escalation comment
-	foundEscalationComment := false
-	for _, call := range client.addCommentCalls {
-		if strings.Contains(call.body, "paused") && strings.Contains(call.body, "Research") {
-			foundEscalationComment = true
-		}
-	}
-	if !foundEscalationComment {
-		t.Error("expected escalation comment to be posted")
-	}
-
-	// pausedDueToRetries should be set
-	itemKey := fmt.Sprintf("%d-%s", 10, "Research")
-	eng.mu.Lock()
-	paused := eng.pausedDueToRetries[itemKey]
-	eng.mu.Unlock()
-	if !paused {
-		t.Error("expected pausedDueToRetries to be set")
-	}
-}
-
-func TestProcessItem_ResetsOnUnpause(t *testing.T) {
-	skipIfNoGit(t)
-	repoDir := initBareRepo(t)
-	wm := NewWorktreeManager(repoDir)
-
-	client := &mockGitHubClient{}
-	claude := &mockClaudeInvoker{
-		invokeFn: func(stage *stages.Stage, issue gh.ProjectItem, newComments []gh.Comment, resume bool, workDir string, modelOverride string) (string, bool, TokenUsage, error) {
-			return "output", false, TokenUsage{}, nil
-		},
-	}
-
-	eng := NewWithDeps(
-		Config{
-			Owner:      "owner",
-			Repo:       "repo",
-			ProjectNum: 1,
-			User:       "testuser",
-			Token:      "token",
-			MaxRetries: 3, // high enough so one retry after unpause doesn't re-escalate
-			Stages:     testStages(),
-		},
-		client,
-		claude,
-		wm,
-	)
-
-	// Simulate a previous escalation: engine had paused this issue after 3 failures
-	itemKey := fmt.Sprintf("%d-%s", 11, "Research")
-	eng.mu.Lock()
-	eng.retryCount[itemKey] = 3
-	eng.pausedDueToRetries[itemKey] = true
-	eng.mu.Unlock()
-
-	board := &gh.ProjectBoard{ProjectID: "PVT_1"}
-	// Item does NOT have fabrik:paused — user has removed it to signal investigation done
-	item := gh.ProjectItem{
-		Number: 11,
-		Title:  "Unpause test",
-		Status: "Research",
-		ItemID: "PVTI_11",
-		Labels: []string{}, // no fabrik:paused
-	}
-
-	if err := eng.processItem(context.Background(), board, item); err != nil {
-		t.Fatalf("processItem: %v", err)
-	}
-
-	// stage:Research:failed should have been removed by clearFailedStage
-	foundRemoval := false
-	for _, call := range client.removeLabelCalls {
-		if call.labelName == "stage:Research:failed" {
-			foundRemoval = true
-		}
-	}
-	if !foundRemoval {
-		t.Error("expected stage:Research:failed label to be removed on unpause")
-	}
-
-	// pausedDueToRetries should be cleared (cleared by clearFailedStage, not re-set since we don't hit limit yet)
-	eng.mu.Lock()
-	stillPaused := eng.pausedDueToRetries[itemKey]
-	eng.mu.Unlock()
-	if stillPaused {
-		t.Error("expected pausedDueToRetries to be cleared after unpause")
-	}
-}
-
-func TestProcessItem_UnlimitedWhenMaxRetriesZero(t *testing.T) {
-	skipIfNoGit(t)
-	repoDir := initBareRepo(t)
-	wm := NewWorktreeManager(repoDir)
-
-	client := &mockGitHubClient{}
-	claude := &mockClaudeInvoker{
-		invokeFn: func(stage *stages.Stage, issue gh.ProjectItem, newComments []gh.Comment, resume bool, workDir string, modelOverride string) (string, bool, TokenUsage, error) {
-			return "output", false, TokenUsage{}, nil
-		},
-	}
-
-	eng := NewWithDeps(
-		Config{
-			Owner:      "owner",
-			Repo:       "repo",
-			ProjectNum: 1,
-			User:       "testuser",
-			Token:      "token",
-			MaxRetries: 0, // unlimited
-			Stages:     testStages(),
-		},
-		client,
-		claude,
-		wm,
-	)
-
-	board := &gh.ProjectBoard{ProjectID: "PVT_1"}
-	item := gh.ProjectItem{Number: 12, Title: "Unlimited retries", Status: "Research", ItemID: "PVTI_12"}
-
-	// Run many times — should never escalate
-	for i := 0; i < 10; i++ {
-		if err := eng.processItem(context.Background(), board, item); err != nil {
-			t.Fatalf("processItem (iteration %d): %v", i, err)
-		}
-	}
-
-	for _, call := range client.addLabelCalls {
-		if call.labelName == "fabrik:paused" {
-			t.Error("should not add fabrik:paused when MaxRetries=0")
-		}
-		if strings.HasSuffix(call.labelName, ":failed") {
-			t.Errorf("should not add failed label when MaxRetries=0, got %q", call.labelName)
-		}
-	}
-
-	// retryCount should remain 0 (not incremented when MaxRetries=0)
-	itemKey := fmt.Sprintf("%d-%s", 12, "Research")
-	eng.mu.Lock()
-	count := eng.retryCount[itemKey]
-	eng.mu.Unlock()
-	if count != 0 {
-		t.Errorf("expected retryCount=0 when MaxRetries=0, got %d", count)
-	}
-}
-
-func TestProcessItem_ClearsRetryCountOnCompletion(t *testing.T) {
-	skipIfNoGit(t)
-	repoDir := initBareRepo(t)
-	wm := NewWorktreeManager(repoDir)
-
-	client := &mockGitHubClient{}
-	claude := &mockClaudeInvoker{
-		invokeFn: func(stage *stages.Stage, issue gh.ProjectItem, newComments []gh.Comment, resume bool, workDir string, modelOverride string) (string, bool, TokenUsage, error) {
-			return "output", true, TokenUsage{}, nil // stage completes successfully
-		},
-	}
-
-	eng := NewWithDeps(
-		Config{
-			Owner:      "owner",
-			Repo:       "repo",
-			ProjectNum: 1,
-			User:       "testuser",
-			Token:      "token",
-			MaxRetries: 3,
-			Stages:     testStages(),
-		},
-		client,
-		claude,
-		wm,
-	)
-
-	// Pre-seed retry state as if previous failures occurred
-	itemKey := fmt.Sprintf("%d-%s", 13, "Research")
-	eng.mu.Lock()
-	eng.retryCount[itemKey] = 2
-	eng.pausedDueToRetries[itemKey] = false
-	eng.mu.Unlock()
-
-	board := &gh.ProjectBoard{ProjectID: "PVT_1"}
-	item := gh.ProjectItem{Number: 13, Title: "Completion test", Status: "Research", ItemID: "PVTI_13"}
-
-	if err := eng.processItem(context.Background(), board, item); err != nil {
-		t.Fatalf("processItem: %v", err)
-	}
-
-	// Both maps should be cleared after successful completion
-	eng.mu.Lock()
-	count := eng.retryCount[itemKey]
-	paused := eng.pausedDueToRetries[itemKey]
-	eng.mu.Unlock()
-
-	if count != 0 {
-		t.Errorf("expected retryCount to be cleared on completion, got %d", count)
-	}
-	if paused {
-		t.Error("expected pausedDueToRetries to be cleared on completion")
-	}
-}
-
-// skipIfNoGit and initBareRepo are defined in worktree_test.go
-
-func testStagesWithCleanup() []*stages.Stage {
-	ss := testStages()
-	return append(ss, &stages.Stage{
-		Name:            "Done",
-		Order:           99,
-		CleanupWorktree: true,
-	})
-}
-
 func TestItemNeedsWork_CleanupStage_NeedsWork(t *testing.T) {
 	eng := testEngine(&mockGitHubClient{}, &mockClaudeInvoker{})
 	eng.cfg.Stages = testStagesWithCleanup()
@@ -1806,167 +507,6 @@ func TestItemNeedsWork_CleanupStage_AlreadyComplete(t *testing.T) {
 	}
 }
 
-func TestProcessItem_CleanupStage_SkipsAlreadyComplete(t *testing.T) {
-	client := &mockGitHubClient{}
-	claude := &mockClaudeInvoker{}
-	eng := testEngine(client, claude)
-	eng.cfg.Stages = testStagesWithCleanup()
-
-	board := &gh.ProjectBoard{ProjectID: "PVT_1"}
-	item := gh.ProjectItem{
-		Number: 1,
-		Title:  "Test",
-		Status: "Done",
-		Labels: []string{"stage:Done:complete"},
-	}
-
-	err := eng.processItem(context.Background(), board, item)
-	if err != nil {
-		t.Fatalf("processItem: %v", err)
-	}
-	if len(client.addLabelCalls) != 0 {
-		t.Errorf("expected no label calls for already-complete cleanup stage, got %d", len(client.addLabelCalls))
-	}
-	if len(claude.calls) != 0 {
-		t.Error("should not invoke claude for cleanup stage")
-	}
-}
-
-func TestProcessItem_CleanupStage_CleanWorktree(t *testing.T) {
-	skipIfNoGit(t)
-	repoDir := initBareRepo(t)
-	wm := NewWorktreeManager(repoDir)
-
-	// Create the worktree first
-	_, err := wm.EnsureWorktree(42, "main", false)
-	if err != nil {
-		t.Fatalf("EnsureWorktree: %v", err)
-	}
-
-	var addedLabel string
-	client := &mockGitHubClient{
-		addLabelToIssueFn: func(owner, repo string, issueNumber int, labelName string) error {
-			addedLabel = labelName
-			return nil
-		},
-	}
-	claude := &mockClaudeInvoker{}
-
-	eng := NewWithDeps(
-		Config{Owner: "owner", Repo: "repo", ProjectNum: 1, User: "testuser", Token: "token",
-			Stages: testStagesWithCleanup()},
-		client, claude, wm,
-	)
-
-	board := &gh.ProjectBoard{ProjectID: "PVT_1"}
-	item := gh.ProjectItem{Number: 42, Title: "Test", Status: "Done", ItemID: "PVTI_42"}
-
-	err = eng.processItem(context.Background(), board, item)
-	if err != nil {
-		t.Fatalf("processItem: %v", err)
-	}
-
-	// Worktree directory should be gone
-	if _, err := os.Stat(wm.WorktreeDir(42)); !os.IsNotExist(err) {
-		t.Error("worktree directory should have been removed")
-	}
-
-	// Completion label should have been added
-	if addedLabel != "stage:Done:complete" {
-		t.Errorf("completion label = %q, want stage:Done:complete", addedLabel)
-	}
-
-	// Should be marked in processedSet
-	eng.mu.Lock()
-	_, ok := eng.processedSet["42-Done"]
-	eng.mu.Unlock()
-	if !ok {
-		t.Error("item should be marked in processedSet after cleanup")
-	}
-
-	// Claude should not have been invoked
-	if len(claude.calls) != 0 {
-		t.Error("claude should not be invoked for cleanup stage")
-	}
-}
-
-func TestProcessItem_CleanupStage_DirtyWorktree(t *testing.T) {
-	skipIfNoGit(t)
-	repoDir := initBareRepo(t)
-	wm := NewWorktreeManager(repoDir)
-
-	// Create the worktree and leave a dirty file
-	wtDir, err := wm.EnsureWorktree(43, "main", false)
-	if err != nil {
-		t.Fatalf("EnsureWorktree: %v", err)
-	}
-	if err := os.WriteFile(filepath.Join(wtDir, "dirty.txt"), []byte("uncommitted"), 0644); err != nil {
-		t.Fatalf("WriteFile: %v", err)
-	}
-
-	client := &mockGitHubClient{}
-	claude := &mockClaudeInvoker{}
-
-	eng := NewWithDeps(
-		Config{Owner: "owner", Repo: "repo", ProjectNum: 1, User: "testuser", Token: "token",
-			Stages: testStagesWithCleanup()},
-		client, claude, wm,
-	)
-
-	board := &gh.ProjectBoard{ProjectID: "PVT_1"}
-	item := gh.ProjectItem{Number: 43, Title: "Test", Status: "Done", ItemID: "PVTI_43"}
-
-	err = eng.processItem(context.Background(), board, item)
-	if err != nil {
-		t.Fatalf("processItem: %v", err)
-	}
-
-	// Worktree directory should still exist (dirty → skip)
-	if _, err := os.Stat(wm.WorktreeDir(43)); os.IsNotExist(err) {
-		t.Error("worktree directory should NOT have been removed for dirty worktree")
-	}
-
-	// No completion label should have been added
-	if len(client.addLabelCalls) != 0 {
-		t.Errorf("expected no label calls for dirty worktree, got %d", len(client.addLabelCalls))
-	}
-}
-
-func TestProcessItem_CleanupStage_NonexistentWorktree(t *testing.T) {
-	skipIfNoGit(t)
-	repoDir := initBareRepo(t)
-	wm := NewWorktreeManager(repoDir)
-	// Don't create the worktree — simulate issue moved to Done before any stage ran
-
-	var addedLabel string
-	client := &mockGitHubClient{
-		addLabelToIssueFn: func(owner, repo string, issueNumber int, labelName string) error {
-			addedLabel = labelName
-			return nil
-		},
-	}
-
-	eng := NewWithDeps(
-		Config{Owner: "owner", Repo: "repo", ProjectNum: 1, User: "testuser", Token: "token",
-			Stages: testStagesWithCleanup()},
-		client, &mockClaudeInvoker{}, wm,
-	)
-
-	board := &gh.ProjectBoard{ProjectID: "PVT_1"}
-	item := gh.ProjectItem{Number: 99, Title: "No Worktree", Status: "Done", ItemID: "PVTI_99"}
-
-	// Should not return error — worktree missing is warn+continue
-	err := eng.processItem(context.Background(), board, item)
-	if err != nil {
-		t.Fatalf("processItem: %v", err)
-	}
-
-	// Completion label should still be added even though worktree didn't exist
-	if addedLabel != "stage:Done:complete" {
-		t.Errorf("completion label = %q, want stage:Done:complete", addedLabel)
-	}
-}
-
 func TestItemNeedsWork_CleanupStage_PausedItem(t *testing.T) {
 	eng := testEngine(&mockGitHubClient{}, &mockClaudeInvoker{})
 	eng.cfg.Stages = testStagesWithCleanup()
@@ -1981,89 +521,3 @@ func TestItemNeedsWork_CleanupStage_PausedItem(t *testing.T) {
 	}
 }
 
-func TestProcessItem_CleanupStage_PRItem(t *testing.T) {
-	// PR items on the board don't have worktrees — cleanup should just apply the label.
-	client := &mockGitHubClient{}
-	claude := &mockClaudeInvoker{}
-	eng := testEngine(client, claude)
-	eng.cfg.Stages = testStagesWithCleanup()
-
-	board := &gh.ProjectBoard{ProjectID: "PVT_1"}
-	item := gh.ProjectItem{
-		Number: 55,
-		Title:  "Some PR",
-		Status: "Done",
-		IsPR:   true,
-		ItemID: "PVTI_55",
-	}
-
-	err := eng.processItem(context.Background(), board, item)
-	if err != nil {
-		t.Fatalf("processItem: %v", err)
-	}
-
-	// Completion label should be applied
-	if len(client.addLabelCalls) != 1 || client.addLabelCalls[0].labelName != "stage:Done:complete" {
-		t.Errorf("expected stage:Done:complete label, got %v", client.addLabelCalls)
-	}
-	if len(claude.calls) != 0 {
-		t.Error("should not invoke claude for cleanup stage PR item")
-	}
-}
-
-func TestProcessItem_CleanupStage_NewCommentsIgnored(t *testing.T) {
-	// New comments on a Done item should not divert to processComments — cleanup runs instead.
-	skipIfNoGit(t)
-	repoDir := initBareRepo(t)
-	wm := NewWorktreeManager(repoDir)
-
-	// Create the worktree
-	_, err := wm.EnsureWorktree(77, "main", false)
-	if err != nil {
-		t.Fatalf("EnsureWorktree: %v", err)
-	}
-
-	var addedLabel string
-	client := &mockGitHubClient{
-		addLabelToIssueFn: func(owner, repo string, issueNumber int, labelName string) error {
-			addedLabel = labelName
-			return nil
-		},
-	}
-	claude := &mockClaudeInvoker{}
-
-	eng := NewWithDeps(
-		Config{Owner: "owner", Repo: "repo", ProjectNum: 1, User: "testuser", Token: "token",
-			Stages: testStagesWithCleanup()},
-		client, claude, wm,
-	)
-
-	// Item has a new (un-rocketed) comment — cleanup should still proceed, not processComments.
-	board := &gh.ProjectBoard{ProjectID: "PVT_1"}
-	item := gh.ProjectItem{
-		Number: 77,
-		Title:  "Test",
-		Status: "Done",
-		ItemID: "PVTI_77",
-		Comments: []gh.Comment{
-			{ID: "C1", Author: "testuser", Body: "please do X"},
-			// No rocket reaction → findNewComments would normally return this
-		},
-	}
-
-	err = eng.processItem(context.Background(), board, item)
-	if err != nil {
-		t.Fatalf("processItem: %v", err)
-	}
-
-	// Worktree should be removed and completion label applied
-	if _, statErr := os.Stat(wm.WorktreeDir(77)); !os.IsNotExist(statErr) {
-		t.Error("worktree directory should have been removed despite new comment")
-	}
-	if addedLabel != "stage:Done:complete" {
-		t.Errorf("completion label = %q, want stage:Done:complete", addedLabel)
-	}
-	if len(claude.calls) != 0 {
-		t.Error("claude should not be invoked for cleanup stage")
-	}
-}
